# openapi_contract.yaml
openapi: 3.0.3
info:
  title: To-Do Backend API
  version: 1.0.0
  description: |
    Formal OpenAPI-style contract for the FastAPI to-do backend.
    This document reflects the currently implemented routes and payload shapes.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /subtasks:
    post:
      tags: [sub-tasks]
      summary: Create sub-task
      operationId: createSubTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubTaskCreate'
      responses:
        '200':
          description: Sub-task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubTaskResponse'
    get:
      tags: [sub-tasks]
      summary: List sub-tasks
      operationId: listSubTasks
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Search'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/SubTaskStatus'
        - in: query
          name: task_id
          schema:
            type: integer
      responses:
        '200':
          description: Sub-task list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubTaskListResponse'
  /subtasks/{sub_task_id}:
    get:
      tags: [sub-tasks]
      summary: Get sub-task by id
      operationId: getSubTaskById
      parameters:
        - $ref: '#/components/parameters/SubTaskIdPath'
      responses:
        '200':
          description: Sub-task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubTaskResponse'
    put:
      tags: [sub-tasks]
      summary: Update sub-task or create approval request
      operationId: updateSubTask
      parameters:
        - $ref: '#/components/parameters/SubTaskIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubTaskUpdate'
      responses:
        '200':
          description: Sub-task response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubTaskResponse'
    delete:
      tags: [sub-tasks]
      summary: Delete sub-task
      operationId: deleteSubTask
      parameters:
        - $ref: '#/components/parameters/SubTaskIdPath'
      responses:
        '200':
          description: Sub-task deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /subtask-update-requests/my:
    get:
      tags: [sub-tasks]
      summary: List my sub-task update requests
      operationId: listMySubTaskUpdateRequests
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Request list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubTaskUpdateRequestListResponse'
  /subtask-update-requests:
    get:
      tags: [sub-tasks]
      summary: List all sub-task update requests (admin only)
      operationId: listAllSubTaskUpdateRequests
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/SubTaskUpdateRequestStatus'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Request list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubTaskUpdateRequestListResponse'
  /subtask-update-requests/{request_id}/approve:
    put:
      tags: [sub-tasks]
      summary: Approve sub-task update request (admin only)
      operationId: approveSubTaskUpdateRequest
      parameters:
        - $ref: '#/components/parameters/RequestIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubTaskUpdateRequestDecision'
      responses:
        '200':
          description: Approved request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubTaskUpdateRequestResponse'
  /subtask-update-requests/{request_id}/reject:
    put:
      tags: [sub-tasks]
      summary: Reject sub-task update request (admin only)
      operationId: rejectSubTaskUpdateRequest
      parameters:
        - $ref: '#/components/parameters/RequestIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubTaskUpdateRequestDecision'
      responses:
        '200':
          description: Rejected request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubTaskUpdateRequestResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      in: query
      name: page
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: page_size
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    Search:
      in: query
      name: search
      required: false
      schema:
        type: string
...
    SubTaskIdPath:
      in: path
      name: sub_task_id
      required: true
      schema:
        type: integer
...
    RequestIdPath:
      in: path
      name: request_id
      required: true
      schema:
        type: integer
...
  schemas:
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    SubTaskStatus:
      type: string
      enum: [complete, not complete]
    SubTaskUpdateRequestStatus:
      type: string
      enum: [pending, approved, rejected]
...
    UserReference:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
...
    SubTaskCreate:
      type: object
      required: [title, description, task_id]
      properties:
        title:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/SubTaskStatus'
        priority:
          type: integer
          minimum: 0
          maximum: 100
          default: 0
        estimated_days:
          type: integer
          minimum: 0
          default: 0
        estimated_hours:
          type: integer
          minimum: 0
          maximum: 23
          default: 0
        actual_days:
          type: integer
          minimum: 0
          default: 0
        actual_hours:
          type: integer
          minimum: 0
          maximum: 23
          default: 0
        task_id:
          type: integer
        assigned_to:
          type: integer
          nullable: true
        assigned_to_username:
          type: string
          nullable: true
    SubTaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/SubTaskStatus'
        priority:
          type: integer
          minimum: 0
          maximum: 100
        estimated_days:
          type: integer
          minimum: 0
        estimated_hours:
          type: integer
          minimum: 0
          maximum: 23
        actual_days:
          type: integer
          minimum: 0
        actual_hours:
          type: integer
          minimum: 0
          maximum: 23
        task_id:
          type: integer
        assigned_to:
          type: integer
          nullable: true
        assigned_to_username:
          type: string
          nullable: true
    SubTaskResponse:
      type: object
      required: [id, title, description, status, priority, estimated_days, estimated_hours, actual_days, actual_hours, created_at, task_id, created_by]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: integer
        estimated_days:
          type: integer
        estimated_hours:
          type: integer
        actual_days:
          type: integer
        actual_hours:
          type: integer
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        task_id:
          type: integer
        created_by:
          $ref: '#/components/schemas/UserReference'
        assigned_to:
          allOf:
            - $ref: '#/components/schemas/UserReference'
          nullable: true
    SubTaskListResponse:
      type: object
      required: [items, total, page, page_size, total_pages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SubTaskResponse'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer
    SubTaskUpdateRequestDecision:
      type: object
      properties:
        comment:
          type: string
          nullable: true
    SubTaskUpdateRequestResponse:
      type: object
      required: [id, sub_task_id, requested_by, status, requested_changes, created_at]
      properties:
        id:
          type: integer
        sub_task_id:
          type: integer
        requested_by:
          $ref: '#/components/schemas/UserReference'
        status:
          $ref: '#/components/schemas/SubTaskUpdateRequestStatus'
        requested_changes:
          type: object
          additionalProperties: true
        review_comment:
          type: string
          nullable: true
        reviewed_by:
          allOf:
            - $ref: '#/components/schemas/UserReference'
          nullable: true
        created_at:
          type: string
          format: date-time
        reviewed_at:
          type: string
          format: date-time
          nullable: true
    SubTaskUpdateRequestListResponse:
      type: object
      required: [items, total, page, page_size, total_pages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SubTaskUpdateRequestResponse'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer
...